+++ doc/version-history.rst
@@ -4,6 +4,19 @@
 Version History
 ===============
 
+v0.6.2
+------
+
+* show forces distribution per quadrant and with XYZ forces off
+* BumpTestTImes class, script to retrieve bump tests runs
+* fix for new qasync library
+* shows heartbeats in VMSGUI
+* Tools for DC accelerometers transformations
+* Disable slew controller changes when slew flag is active
+* improved acceleration - forces fits
+* improved unit displays
+* improved command line processing
+
 v0.6.1
 ------